Next step of config overhaul, colocate all config with their functionality as much as possible
