+++ src/instrumentation/queue.ts
@@ -1,8 +1,9 @@
 import { trace, SpanOptions, SpanKind, Attributes, Exception, context as api_context } from '@opentelemetry/api'
-import { Initialiser, setConfig } from '../config'
+import { setConfig } from '../config'
 import { exportSpans, proxyExecutionContext } from './common'
 import { instrumentEnv } from './env'
 import { unwrap, wrap } from './wrap'
+import { Initialiser } from '../types'
 
 type QueueHandler = ExportedHandlerQueueHandler<unknown, unknown>
 export type QueueHandlerArgs = Parameters<QueueHandler>
