+++ src/sdk.ts
@@ -6,11 +6,11 @@ import { Resource } from '@opentelemetry/resources'
 import { SemanticResourceAttributes } from '@opentelemetry/semantic-conventions'
 import { SpanExporter } from '@opentelemetry/sdk-trace-base'
 
-import { Initialiser, Trigger, TraceConfig, ResolvedTraceConfig, ExporterConfig } from './config'
 import { OTLPExporter } from './exporter'
 import { WorkerTracerProvider } from './provider'
 import { isHeadSampled, isRootErrorSpan, multiTailSampler } from './sampling'
 import { BatchTraceSpanProcessor } from './spanprocessor'
+import { Trigger, TraceConfig, ResolvedTraceConfig, ExporterConfig, Initialiser } from './types'
 import { createFetchHandler, instrumentGlobalFetch } from './instrumentation/fetch'
 import { instrumentGlobalCache } from './instrumentation/cache'
 import { createQueueHandler } from './instrumentation/queue'
